refactor account creation
